refactor: new structure for routes
