+++ app/routes/__docs/$chapter.$slug.tsx
