+++ app/routes/__docs/$slug.tsx
